chore(ws-core): more refactoring work
