+++ packages/graphql-server-reactive-core/src/index.ts
@@ -1,3 +1,5 @@
 export * from 'graphql-server-core';
 export { Observable, IObservable, Observer, Subscription } from './Observable';
 export { runQueryReactive, ReactiveExecuteFunction, ReactiveQueryOptions } from './runQueryReactive';
+export { ReactiveGraphQLEngine, ReactiveGraphQLOptions } from './reactiveOptions';
+export { RequestsManager, ReactiveRequest, ReactiveMessage } from './RequestsManager';
