+++ packages/graphql-server-reactive-core/src/reactiveOptions.ts
@@ -1,0 +1,10 @@
+import { GraphQLOptions } from 'graphql-server-core';
+import { ReactiveExecuteFunction } from './runQueryReactive';
+
+export interface ReactiveGraphQLEngine {
+  executeReactive: ReactiveExecuteFunction;
+}
+
+export interface ReactiveGraphQLOptions extends GraphQLOptions {
+  engine: ReactiveGraphQLEngine;
+}
