+++ packages/graphql-server-ws/package.json
@@ -25,8 +25,7 @@
   },
   "homepage": "https://github.com/apollostack/graphql-server#readme",
   "dependencies": {
-    "graphql-server-reactive-core": "^0.6.0",
-    "observable-diff-operator": "^0.1.1"
+    "graphql-server-reactive-core": "^0.6.0"
   },
   "devDependencies": {
     "ws": "^2.0.2",
