--- packages/graphql-server-ws/src/interfaces.ts
@@ -1,35 +1,0 @@
-import { GraphQLOptions, ReactiveExecuteFunction } from 'graphql-server-reactive-core';
-import { graphql, GraphQLSchema, ExecutionResult } from 'graphql';
-import * as WebSocket from 'ws';
-
-export interface WSRequest {
-  id: number; // Per socket increasing number
-  action: 'request' | 'cancel';
-  query?: string; // GraphQL Printed Query.
-  variables?: any; // GraphQL variables.
-  operationName?: string; // GraphQL operationName
-}
-
-export interface WSMessageParams {
-  requestParams: WSRequest;
-  graphqlOptions?: WSGraphQLOptions;
-  flags: {
-    binary: boolean;
-  };
-}
-
-export interface WSGraphQLEngine {
-  executeReactive: ReactiveExecuteFunction;
-}
-
-export interface WSGraphQLOptions extends GraphQLOptions {
-  engine: WSGraphQLEngine;
-}
-
-export interface WSGraphQLOptionsFunction {
-  (ws: WebSocket): WSGraphQLOptions | Promise<WSGraphQLOptions>;
-}
-
-export interface WSHandler {
-  (ws: WebSocket): void;
-}
